Add `bundle install` step to the getting started guide
When I was following the guide I had to do this step manually, and
those who are newer to the Ruby ecosystem might not know this
command is required before the `bundle exec hanami dev` command.

I hope this is helpful! Happy to change as needed! (Or maybe I did
something wrong?)

Best,
Pat